+++ src/Controller/ClasificationController.php
@@ -10,33 +10,34 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
-/**
- * @Route("{_locale}/clasification")
- * @isGranted("ROLE_ADMIN")
- */
+#[Route(path: '{_locale}/clasification')]
+#[IsGranted('ROLE_ADMIN')]
 class ClasificationController extends AbstractController
 {
-    /**
-     * @Route("/", name="app_clasification_index", methods={"GET"})
-     */
+
+    public function __construct(private readonly TranslatorInterface $translator)
+    {       
+    }
+
+    #[Route(path: '/', name: 'app_clasification_index', methods: ['GET'])]
     public function index(Request $request, ClasificationRepository $clasificationRepository): Response
     {
-        $ajax = $request->get('ajax') !== null ? $request->get('ajax') : "false";
+        $ajax = $request->get('ajax') ?? "false";
         $form = $this->createForm(ClasificationType::class);
         $template = $ajax === "true" ? '_list.html.twig' : 'index.html.twig';
         return $this->render("clasification/$template", [
             'clasifications' => $clasificationRepository->findAll(),
-            'form' => $form->createView(),
+            'form' => $form,
         ]);
     }
 
     /**
      * Creates or updates a Activity
-     * 
-     * @Route("/new", name="app_clasification_save", methods={"GET","POST"})
      */
+    #[Route(path: '/new', name: 'app_clasification_save', methods: ['GET', 'POST'])]
     public function createOrSave(Request $request, ClasificationRepository $repo, EntityManagerInterface $em): Response
     {
         $clasification = new Clasification();
@@ -54,7 +55,7 @@ class ClasificationController extends AbstractController
             $em->flush();
 
             if ($request->isXmlHttpRequest()) {
-                return new Response(null, 204);
+                return new Response(null, Response::HTTP_NO_CONTENT);
             }
             return $this->redirectToRoute('app_clasification_index');
         }
@@ -62,13 +63,11 @@ class ClasificationController extends AbstractController
         $template = $request->isXmlHttpRequest() ? '_form.html.twig' : 'new.html.twig';
         return $this->render('clasification/' . $template, [
             'holiday' => $clasification,
-            'form' => $form->createView(),
+            'form' => $form,
         ], new Response(null, $form->isSubmitted() && !$form->isValid() ? 422 : 200,));
     }
 
-    /**
-     * @Route("/{id}", name="app_clasification_show", methods={"GET"})
-     */
+    #[Route(path: '/{id}', name: 'app_clasification_show', methods: ['GET'])]
     public function show(Request $request, Clasification $clasification, EntityManagerInterface $entityManager): Response
     {
         $form = $this->createForm(ClasificationType::class, $clasification, [
@@ -85,14 +84,12 @@ class ClasificationController extends AbstractController
         $template = $request->isXmlHttpRequest() ? '_form.html.twig' : 'edit.html.twig';
         return $this->render('clasification/' . $template, [
             'clasification' => $clasification,
-            'form' => $form->createView(),
+            'form' => $form,
             'readonly' => true,
-        ], new Response(null, $form->isSubmitted() && !$form->isValid() ? 422 : 200,));
+        ], new Response(null, $form->isSubmitted() && !$form->isValid() ? Response::HTTP_UNPROCESSABLE_ENTITY : Response::HTTP_OK,));
     }
 
-    /**
-     * @Route("/{id}/edit", name="app_clasification_edit", methods={"GET", "POST"})
-     */
+    #[Route(path: '/{id}/edit', name: 'app_clasification_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Clasification $id, EntityManagerInterface $entityManager): Response
     {
         $form = $this->createForm(ClasificationType::class, $id, [
@@ -109,16 +106,17 @@ class ClasificationController extends AbstractController
         $template = $request->isXmlHttpRequest() ? '_form.html.twig' : 'edit.html.twig';
         return $this->render('clasification/' . $template, [
             'clasification' => $id,
-            'form' => $form->createView(),
+            'form' => $form,
             'readonly' => false
-        ], new Response(null, $form->isSubmitted() && !$form->isValid() ? 422 : 200,));
+        ], new Response(null, $form->isSubmitted() && !$form->isValid() ? Response::HTTP_UNPROCESSABLE_ENTITY : Response::HTTP_OK,));
     }
 
-    /**
-     * @Route("/{id}", name="app_clasification_delete", methods={"POST","DELETE"})
-     */
+    #[Route(path: '/{id}', name: 'app_clasification_delete', methods: ['POST', 'DELETE'])]
     public function delete(Request $request, Clasification $id, EntityManagerInterface $entityManager): Response
     {
+        if ( !$this->canBeDeleted($id) ) {
+            return new Response($this->translator->trans('error.clasificationHasActivities'),Response::HTTP_NOT_ACCEPTABLE);
+        }
         if ($this->isCsrfTokenValid('delete'.$id->getId(), $request->get('_token'))) {
             $entityManager->remove($id);
             $entityManager->flush();
@@ -128,7 +126,15 @@ class ClasificationController extends AbstractController
                 return new Response(null, Response::HTTP_NO_CONTENT);
             }
         } else {
-            return new Response('messages.invalidCsrfToken', 422);
+            return new Response('messages.invalidCsrfToken', Response::HTTP_UNPROCESSABLE_ENTITY);
         }
+    }
+
+    private function canBeDeleted(Clasification $clasification): bool {
+        $activities = $clasification->getActivitys();
+        if (count($activities) > 0) {
+            return false;
+        }
+        return true;
     }
 }
